Organize user under domain package
