+++ src/main/java/org/primeframework/mvc/sampleapp/action/LoginAction.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 import com.google.inject.Inject;
 import org.primeframework.mvc.action.annotation.Action;
 import org.primeframework.mvc.action.result.annotation.ReexecuteSavedRequest;
-import org.primeframework.mvc.sampleapp.session.OurUser;
+import org.primeframework.mvc.sampleapp.domain.OurUser;
 import org.primeframework.mvc.security.UserLoginSecurityContext;
 
 @Action
