+++ src/main/java/org/primeframework/mvc/sampleapp/action/PageThatRequiresAuthAction.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import com.google.inject.Inject;
 import org.primeframework.mvc.action.annotation.Action;
-import org.primeframework.mvc.sampleapp.session.OurUser;
+import org.primeframework.mvc.sampleapp.domain.OurUser;
 import org.primeframework.mvc.security.UserLoginSecurityContext;
 
 @Action(requiresAuthentication = true)
