+++ src/main/java/org/primeframework/mvc/sampleapp/domain/OurUser.java
@@ -1,4 +1,4 @@
-package org.primeframework.mvc.sampleapp.session;
+package org.primeframework.mvc.sampleapp.domain;
 
 import java.util.UUID;
