+++ src/main/java/org/primeframework/mvc/sampleapp/session/ExampleLoginContext.java
@@ -10,6 +10,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
 import io.fusionauth.http.server.HTTPRequest;
 import io.fusionauth.http.server.HTTPResponse;
+import org.primeframework.mvc.sampleapp.domain.OurUser;
 import org.primeframework.mvc.security.BaseUserIdCookieSecurityContext;
 import org.primeframework.mvc.security.Encryptor;
 import org.primeframework.mvc.security.UserIdSessionContext;
